Checkout all projects with the same Change-Id
If there's more than one project with a change in review with the
same Change-Id, check it out as well. This way you can easily
have dependencies between changes and make the build bots aware
of it.

Reviewed-on: http://review.couchbase.org/59193
Tested-by: Will Gardner <will.gardner@couchbase.com>

+++ scripts/jenkins/commit_validation/fetch_project.sh
@@ -10,8 +10,17 @@ PROJECT=$1
 PROJECT_PATH=$2
 REFSPEC=$3
 
+if [ -z "$GERRIT_HOST" ]; then
+    echo "Error: Required environment variable 'GERRIT_HOST' not set."
+    exit 1
+fi
+if [ -z "$GERRIT_PORT" ]; then
+    echo "Error: Required environment variable 'GERRIT_PORT' not set."
+    exit 2
+fi
+
 cd $PROJECT_PATH
 git reset --hard HEAD
-git fetch ssh://review.couchbase.org:29418/$PROJECT $REFSPEC
+git fetch ssh://$GERRIT_HOST:$GERRIT_PORT/$PROJECT $REFSPEC
 git checkout FETCH_HEAD
 cd ..
